quanto-gastei-backend/requirements.txt
#Dependencias do projeto
flask
pymongo
python-dotenv
azure-storage-blob
